requirements.txt
aiodns==1.2.0
aiohttp==3.7.4
altgraph==0.17.2
anytree==2.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
ascii-art-cli==1.4.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.0.5
async-timeout==3.0.1
attrs==18.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.7.1
bipwallet==1.0.2
bit==0.8.0
bitcoinaddress==0.1.5
bitcoinlib==0.6.7
bleach==4.1.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
coincurve==18.0.0
colorama==0.4.4
contourpy==1.0.6
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-embed-video==1.4.8
easy-thumbnails==2.8.5
ecdsa==0.18.0
entrypoints==0.4
executing==0.8.3
fake-useragent==1.1.1
Flask==2.1.2
fonttools==4.38.0
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
grequests==0.6.0
idna==2.8
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso-639==0.4.5
iso3166==0.9
isodate==0.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.11.0
keyboard==0.13.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
maxminddb==2.2.0
mistune==0.8.4
mnemonic==0.20
MouseInfo==0.1.3
multidict==4.5.2
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
nickname-generator==1.0.1
notebook==6.4.10
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandocfilters==1.5.0
paperclip==2.7.2
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==8.4.0
playwright==1.31.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
py-temp-mails-api==1.38
PyAutoGUI==0.9.53
pycares==4.3.0
pycparser==2.21
pycryptodome==3.16.0
pyee==9.0.4
PyGetWindow==0.0.9
Pygments==2.11.2
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.6.8
pyTelegramBotAPI==4.5.1
python-dateutil==2.8.2
python-magic==0.4.27
pytweening==1.0.4
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
random-password-generator==2.2.0
randomtimestamp==2.2
requests==2.27.1
selenium==3.141.0
Send2Trash==1.8.0
six==1.12.0
soupsieve==1.7.3
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.2.0
streamlink==1.0.0
tabulate==0.9.0
tempMail2==1.0.2
terminado==0.13.3
testpath==0.6.0
TikTokApi==5.2.2
tornado==6.1
traitlets==5.1.1
two1==3.10.9
typing==3.6.6
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.5
vk-api==11.9.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.55.0
Werkzeug==2.1.2
widgetsnbextension==3.6.0
yarl==1.3.0
zope.event==4.6
zope.interface==5.5.2
